Make HitResult as abtract
