+++ Source/HelixToolkit.Wpf/SelectionCommands/CombinedSelectionCommand.cs
@@ -122,7 +122,7 @@ public sealed class CombinedSelectionCommand : SelectionCommand
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         void HandlePointSelection()
         {
-            IList<Viewport3DHelper.HitResult>? res = this.Viewport.FindHits(this.selectionRect.Location) ?? new List<Viewport3DHelper.HitResult>();
+            IList<PointHitResult>? res = this.Viewport.FindHits(this.selectionRect.Location) ?? new List<PointHitResult>();
             var selectedModels = res.Select(hit => hit.Model).ToList();
             this.OnModelsSelected(new ModelsSelectedByPointEventArgs(selectedModels, this.selectionRect.Location));
             var selectedVisuals = res.Select(hit => hit.Visual).ToList();
@@ -131,7 +131,7 @@ public sealed class CombinedSelectionCommand : SelectionCommand
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         void HandleRectangleSelection()
         {
-            IEnumerable<Viewport3DHelper.RectangleHitResult> res = this.Viewport.FindHits(this.selectionRect, this.SelectionHitMode);
+            IEnumerable<RectangleHitResult> res = this.Viewport.FindHits(this.selectionRect, this.SelectionHitMode);
             var selectedModels = res.Select(hit => hit.Model).ToList();
             this.OnModelsSelected(new ModelsSelectedByRectangleEventArgs(selectedModels, this.selectionRect));
             var selectedVisuals = res.Select(hit => hit.Visual).ToList();
